@@ -25,7 +25,7 @@ classifiers = [
     "Programming Language :: C",
     "Topic :: Software Development :: Version Control",
 ]
-requires-python = ">=3.9,<3.15"
+requires-python = ">=3.10,<3.15"
 dependencies = [
     "configobj",
     "fastbencode",
