@@ -20,7 +20,6 @@
 import posixpath
 import stat
 from collections.abc import Iterable, Iterator
-from typing import Optional
 
 from dulwich.object_store import BaseObjectStore
 from dulwich.objects import ZERO_SHA, Blob, Commit, ObjectID, ShaFile, Tree, sha_to_hex
@@ -798,7 +797,7 @@ class BazaarObjectStore(BaseObjectStore):
         self,
         haves: Iterable[ObjectID],
         wants: Iterable[ObjectID],
-        shallow: Optional[Iterable[ObjectID]] = None,
+        shallow: Iterable[ObjectID] | None = None,
         progress=None,
         get_tagged=None,
         get_parents=None,
